Mise a jour du code pour les modification BDD + Gestion du prix total decimal lors de la reservation.

+++ MRLoca/Dao/DaoHebergement.cs
@@ -199,7 +199,7 @@ namespace MRLoca.Dao
                 }
 
 
-                hebergement.PrixDeBase = Convert.ToInt32(base.sqlDataReader["PrixDeBase"]);
+                hebergement.PrixDeBase = Convert.ToDecimal(base.sqlDataReader["PrixDeBase"]);
 
                 if(Convert.ToString(base.sqlDataReader["IdAdresse"]) != "" && Convert.ToString(base.sqlDataReader["NomAdresse"]) != "")
                 {
